exposed wireframe, culling, blending and depthtest state functions; fixed a crash hapening when closing the console window first fixed Matrix::Transpose and Matrix::GetTransposed fixed perspective projection and view matrices issues (kinda hacky... need to look into column major issue in GLSL) fixed FileLoaderManager and RM cleanup hack for reducing character overlapping in output text renamed log functions to be shorter replaced std::shared_ptr with a shorter alias (sp) added a lumberjack model instead of the blender monkey (by 3DAssetForge, turbosquid) exposed fov in construct for main freecam

+++ src/LLAPI/Utils/RendererUtils.h
@@ -7,7 +7,7 @@ namespace oakvr
 {
 	namespace render
 	{
-		enum class PrimitiveTopology : uint8_t
+		enum class PrimitiveTopology : uint32_t
 		{
 			ePT_First = 0,
 			ePT_PointList = ePT_First,
@@ -20,7 +20,7 @@ namespace oakvr
 
 		struct VertexElementDescriptor
 		{			
-			enum class Semantic : uint8_t
+			enum class Semantic : uint32_t
 			{
 				position,
 				tex_coord,
